rand_distr: add module student_t
